refactor(vscode-webui): merge built-in and custom agent sections
Combine built-in and custom agent settings sections into a single unified agent section component to simplify the settings UI hierarchy.
